Crate eq_oracle

Equilibrium Oracle Pallet

Equilibrium's Oracle Pallet is a Substrate module for receiving external prices information, aggregating it into currency prices and using them in Substrate

Modules

crypto
Module for signing operations

Structs

DataPoint
Struct for storing added currency price data from one source

GenesisConfig
Genesis config for the module, allow to build genesis storage.

Module

PricePoint
Struct for storing aggregated currency price data

Enums

Call
Dispatchable calls.

Error
Pallet's errors

PriceSource
Source of received price data points

RawEvent
Events for this module.

Constants

KEY_TYPE

Traits

OnNewPrice
Newly added price points processing logic

PriceGetter
Interface for getting current system price of a currency

Trait
Substrate pallet configuration trait

Type Definitions

Event
RawEvent specialized for the configuration Trait