Trait eq_oracle::Trait

pub trait Trait: CreateSignedTransaction<Call<Self>> {
    type AuthorityId: AppCrypto<Self::Public, Self::Signature>;
    type Event: From<Event<Self>> + Into<Self::Event>;
    type Call: From<Call<Self>>;
    type UnixTime: UnixTime;
    type Whitelist: CheckWhitelisted<Self::AccountId>;
    type MedianPriceTimeout: Get<u64>;
    type PriceTimeout: Get<u64>;
    type OnNewPrice: OnNewPrice;
}

Substrate pallet configuration trait

Associated Types

type AuthorityId: AppCrypto<Self::Public, Self::Signature>

type Event: From<Event<Self>> + Into<Self::Event>

type Call: From<Call<Self>>

type UnixTime: UnixTime

Timestamp provider

type Whitelist: CheckWhitelisted<Self::AccountId>

Whitelist checks for price setters

type MedianPriceTimeout: Get<u64>

Pallet setting representing amount of time for which price median is valid

type PriceTimeout: Get<u64>

Pallet setting representing amount of time for which price point is valid

type OnNewPrice: OnNewPrice

Newly added price points processing logic

Loading content...